requirements.txt
flask
flask-cors
python-dotenv
requests
sentence-transformers
PyMuPDF
faiss-cpu
langchain
scikit-learn
